Changed CiscoIOSGeneric script to use centralised cleanup script
Previous version had changes to multiple files to clean up unwanted
"ntp clock-period" lines. This version replaces all that with a single
script. This means only one place to change if we want to strip out
other lines.

It also means the only changes are to edit the
CiscoIOSGeneric_Config_Backup_Builder*.xml files, and to add one new
script. This is much cleaner than the previous version, with fewer
changes to the HP-provided files.